Ponder Answer Comments: indent textarea as Comments
Summary:
In a Ponder Answer, when adding a Comment, the textarea indentation was a
a little bit unintuitive.

After this change, the textarea is aligned as a Comment:

| Before    | After     |
|-----------|-----------|
| {F286220} | {F286221} |

Other changes are welcome but this seems to me a good minimum.

I also hope this could help not to confuse the Comment field with the Answer field.

Closes T15350

Test Plan:
- create a Ponder Question
- create an Answer
- look at the Comments input field that now should be more nicely indented
- the normal Response field is not changed at all
- the mobile view is not changed at all
- click on "Pin Form On Screen" and it must be as before

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: speck, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T15350

Differential Revision: https://we.phorge.it/D25192