Replacing the deprecated -moz-outline-style with outline-style
Summary:
Replacing the deprecated -moz-outline-style with outline-style: none.

According to "Can I Use", this change is supported in whatever non-ridiculously-old browser web in the universe,
like Firefox 2 and Chrome 4 and Internet Explorer 8.

https://caniuse.com/outline

Ref T15585

Test Plan: The outline in the <a> tags should look as usual, that is, none.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15585

Differential Revision: https://we.phorge.it/D25416