The feed "created this task" should be the first one
Summary:
Create a Maniphest task, Resolved, High priority.

Before this change, this was the feed in the object page:

- a.lincoln closed this task as Resolved.
- a.lincoln triaged this task as High priority.
- a.lincoln __created this task__.

After this change, this is the feed:

- a.lincoln __created this task__.
- a.lincoln closed this task as Resolved.
- a.lincoln triaged this task as High priority.

This also improves the situation in the /feed/ page. So you see the creation now.

Ref T15816

Test Plan:
Create an object in any mentioned application below, setting all fields.

Then, check the history of the object itself. Check also the /feed/ page.
Check that the creation action is now listed first.

This change improves the situation in:

- Maniphest
- Phriction

This change has no impact on these that seem already OK:

- Calendar
- Dashboard
- Dashboard Panel
- Diffusion
- Macro
- Paste
- Ponder
- Project
- Pholio

This change has no impact on these that seem still in the wrong order:

- Differential

The above situation in Differential is noted, but it's unrelated,
since it has a different legacy structure and it deserves more triaging.

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T15816

Differential Revision: https://we.phorge.it/D25626